jobs-abc/requirements.txt
Django==2.1.7
pytz==2018.9
whitenoise==4.1.2
gunicorn
psycopg2
djangorestframework
pydevd
